Implementando tela de confirmação de deleção

+++ lib/screens/home_screen/widgets/journal_card.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_webapi_first_course/helpers/weekday.dart';
 import 'package:flutter_webapi_first_course/models/journal.dart';
+import 'package:flutter_webapi_first_course/screens/common/confirmation_dialog.dart';
 import 'package:flutter_webapi_first_course/services/journal_service.dart';
 import 'package:uuid/uuid.dart';
 
@@ -9,7 +10,11 @@ class JournalCard extends StatelessWidget {
   final DateTime showedDate;
   final Function refreshFunction;
 
-  const JournalCard({Key? key, this.journal, required this.showedDate, required this.refreshFunction})
+  const JournalCard(
+      {Key? key,
+      this.journal,
+      required this.showedDate,
+      required this.refreshFunction})
       : super(key: key);
 
   @override
@@ -79,7 +84,11 @@ class JournalCard extends StatelessWidget {
                   ),
                 ),
               ),
-              IconButton(onPressed: (){removeJournal(context);}, icon: const Icon(Icons.delete))
+              IconButton(
+                  onPressed: () {
+                    removeJournal(context);
+                  },
+                  icon: const Icon(Icons.delete))
             ],
           ),
         ),
@@ -111,10 +120,10 @@ class JournalCard extends StatelessWidget {
 
     Map<String, dynamic> map = {};
 
-    if (journal != null){
+    if (journal != null) {
       innerJournal = journal;
       map['is_editing'] = false;
-    }else {
+    } else {
       map['is_editing'] = true;
     }
 
@@ -130,7 +139,7 @@ class JournalCard extends StatelessWidget {
         ScaffoldMessenger.of(context).showSnackBar(
           const SnackBar(content: Text("Registro realizado com sucesso!")),
         );
-      }else{
+      } else {
         ScaffoldMessenger.of(context).showSnackBar(
           const SnackBar(content: Text("Houve falha ao registrar!")),
         );
@@ -138,19 +147,31 @@ class JournalCard extends StatelessWidget {
     });
   }
 
-  removeJournal (BuildContext context) {
+  removeJournal(BuildContext context) {
     JournalService service = JournalService();
 
-    if (journal != null){
-      service.delete(journal!.id).then((value) {
-        if (value) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text("Removido com sucesso!"),
-            ), // SnackBar
-          );
+    if (journal != null) {
 
-          refreshFunction();
+      showConfirmationDialog(
+        context,
+        content:
+            "Deseja realmente remover a atividade do dia ${WeekDay(journal!.createdAt)}?",
+        affirmativeOption: "Remover"
+      ).then((value) {
+        if (value != null){
+          if (value){
+            service.delete(journal!.id).then((value) {
+              if (value) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  const SnackBar(
+                    content: Text("Removido com sucesso!"),
+                  ), // SnackBar
+                );
+
+                refreshFunction();
+              }
+            });
+          }
         }
       });
     }
